chore: switch `syncCoreSets` to use shared `listOutdated` impl
